requirements.txt
pyESDL==21.4.1

Werkzeug==0.16.1
Flask==1.1.4
Flask-Cors==3.0.8
flask-restplus==0.13.0
